ISSUE #3725 - make navy modal header into reusable component
